Refactored settings store into a separate file
